third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-order-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'order' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.orderX.baseVal is 2
PASS convolveMatrixElement.orderY.baseVal is 2
PASS convolveMatrixElement.orderX.baseVal is 3
PASS convolveMatrixElement.orderY.baseVal is 3
PASS successfullyParsed is true

TEST COMPLETE
